requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
google-generativeai==0.3.2
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.0
pydantic-settings==2.1.0
httpx==0.27.0
mercadopago